Division hotfix

+++ src/AppVeyorTest/Class1.cs
@@ -6,6 +6,6 @@ namespace AppVeyorTest
     {
         public int Add(int x, int y) => x + y;
         public int Subtract(int x, int y) => x - y;
-        public double Divide(int x, int y) => x / y;
+        public double Divide(int x, int y) => (double)x / (double)y;
     }
 }
